+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.13
+
+- Fixed availability of "Refresh from Disk" for untitled Leo documents.
+- Fixed default base directory for external files of untitled Leo documents.
+- Fixed 'import txt file' usage of relative directories.
+- Fixed node insertion logic (after node instead of as last child) for the CWEB importer.
+
 ## 1.0.12
 
 - Fixed scrolling to reveal cursor on selected line for gotoGlobalLine related commands.